+++ mosaicod/bin/mosaicod/src/command/run.rs
@@ -3,7 +3,6 @@ use clap::Args;
 use mosaicod_core::{self as core, error::PublicResult as Result, params};
 use mosaicod_db as db;
 use mosaicod_grpc as grpc;
-use mosaicod_grpc_flight as grpc_flight;
 use signal_hook::{consts::SIGINT, iterator::Signals};
 use std::thread;
 use tracing::{debug, info};
@@ -33,8 +32,8 @@ pub struct Run {
     pub api_key: bool,
 }
 
-fn tls_config() -> grpc_flight::TlsConfig {
-    grpc_flight::TlsConfig {
+fn tls_config() -> grpc::TlsConfig {
+    grpc::TlsConfig {
         certificate_file: params::params().tls_certificate_file.value.clone().into(),
         private_key_file: params::params().tls_private_key_file.value.clone().into(),
     }
@@ -74,15 +73,15 @@ pub fn run(args: Run, json_format: bool) -> Result<()> {
     let mut server = grpc::Server::new(host, args.port, store, db);
 
     if args.api_key {
-        server.flight_config.enable_api_key_management();
+        server.options.enable_api_key_management();
     }
 
     if args.tls {
-        server.flight_config.tls(tls_config());
+        server.options.tls(tls_config());
     }
 
     if args.gzip {
-        server.flight_config.gzip(true);
+        server.options.gzip(true);
     }
 
     // (cabba) NOTE: maybe we need to return a more specific error ?
