+++ mosaicod/crates/mosaicod-grpc/Cargo.toml
@@ -15,7 +15,11 @@ mosaicod-grpc-flight = { workspace = true }
 mosaicod-core = { workspace = true }
 mosaicod-db = { workspace = true }
 mosaicod-store = { workspace = true }
+mosaicod-ext = { workspace = true }
 mosaicod-task = { workspace = true }
 
+arrow-flight = { workspace = true }
+tonic = { workspace = true }
+tower = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
